+++ CHANGELOG.md
@@ -9,5 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Introduced a brand new CLI interface
 - The preferred method of installation is to use the Zephir PHAR
   which can be downloaded from the most recent Github Release
+- Zephir development mode will be enabled silently if your PHP binary was compiled in
+  a debug configuration [#1520](https://github.com/phalcon/zephir/issues/1520)
 
 [Unreleased]: https://github.com/phalcon/zephir/compare/0.11.3...HEAD